Add tests for the rating adapter + Modify the method in database

+++ app/src/main/java/com/github/sdpteam15/polyevents/model/entity/Rating.kt
@@ -2,7 +2,7 @@ package com.github.sdpteam15.polyevents.model.entity
 
 data class Rating(
     val ratingId: String? = null,
-    val rate: Double? = null,
+    val rate: Float? = null,
     val feedback: String? = null,
     val eventId: String? = null,
     val userId: String? = null
